"Small refactoring". 🙈
