+++ cmd/server/token-cmd.go
@@ -7,8 +7,7 @@ import (
 
 	apiv2 "github.com/metal-stack/api/go/metalstack/api/v2"
 	"github.com/metal-stack/metal-apiserver/pkg/certs"
-	"github.com/metal-stack/metal-apiserver/pkg/service/api/token"
-	tokencommon "github.com/metal-stack/metal-apiserver/pkg/token"
+	"github.com/metal-stack/metal-apiserver/pkg/token"
 	"github.com/urfave/cli/v2"
 	"google.golang.org/protobuf/types/known/durationpb"
 )
@@ -79,18 +78,6 @@ func newTokenCmd() *cli.Command {
 				return err
 			}
 
-			tokenStore := tokencommon.NewRedisStore(tokenRedisClient)
-			certStore := certs.NewRedisStore(&certs.Config{
-				RedisClient: tokenRedisClient,
-			})
-
-			tokenService := token.New(token.Config{
-				Log:        log,
-				TokenStore: tokenStore,
-				CertStore:  certStore,
-				Issuer:     ctx.String(serverHttpUrlFlag.Name),
-			})
-
 			var permissions []*apiv2.MethodPermission
 			for _, m := range ctx.StringSlice(tokenPermissionsFlag.Name) {
 				project, semicolonSeparatedMethods, ok := strings.Cut(m, "=")
@@ -149,7 +136,13 @@ func newTokenCmd() *cli.Command {
 				return fmt.Errorf("token subject cannot be empty")
 			}
 
-			resp, err := tokenService.CreateApiTokenWithoutPermissionCheck(ctx.Context, subject, &apiv2.TokenServiceCreateRequest{
+			resp, err := token.NewWithoutPermissionCheck(&token.TokenWithoutPermissionCheckConfig{
+				Tokens: token.NewRedisStore(tokenRedisClient),
+				Certs: certs.NewRedisStore(&certs.Config{
+					RedisClient: tokenRedisClient,
+				}),
+				Issuer: ctx.String(serverHttpUrlFlag.Name),
+			}).CreateApiTokenWithoutPermissionCheck(ctx.Context, subject, &apiv2.TokenServiceCreateRequest{
 				Description:  ctx.String(tokenDescriptionFlag.Name),
 				Expires:      durationpb.New(ctx.Duration(tokenExpirationFlag.Name)),
 				ProjectRoles: projectRoles,
