+++ pkg/certs/certs-store_test.go
@@ -6,7 +6,6 @@ import (
 	"github.com/alicebob/miniredis/v2"
 	"github.com/lestrrat-go/jwx/v3/jwk"
 	"github.com/metal-stack/metal-apiserver/pkg/certs"
-	"github.com/metal-stack/metal-apiserver/pkg/token"
 	"github.com/redis/go-redis/v9"
 	"github.com/stretchr/testify/require"
 )
@@ -18,7 +17,7 @@ func Test_redisStore(t *testing.T) {
 		s     = miniredis.RunT(t)
 		c     = redis.NewClient(&redis.Options{Addr: s.Addr()})
 		store = certs.NewRedisStore(&certs.Config{
-			RenewCertBeforeExpiration: new(4 * token.MaxExpiration),
+			RenewCertBeforeExpiration: new(4 * certs.MaxTokenExpiration),
 			RedisClient:               c,
 		})
 	)
