+++ pkg/service/admin/token/token-service_test.go
@@ -1,0 +1,360 @@
+package admin
+
+import (
+	"context"
+	"log/slog"
+	"testing"
+
+	"buf.build/go/protovalidate"
+	"github.com/alicebob/miniredis/v2"
+	"github.com/google/go-cmp/cmp"
+	adminv2 "github.com/metal-stack/api/go/metalstack/admin/v2"
+	apiv2 "github.com/metal-stack/api/go/metalstack/api/v2"
+	"github.com/metal-stack/metal-apiserver/pkg/certs"
+	"github.com/metal-stack/metal-apiserver/pkg/repository/api"
+	"github.com/metal-stack/metal-apiserver/pkg/token"
+	"github.com/redis/go-redis/v9"
+	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/testing/protocmp"
+)
+
+func Test_List(t *testing.T) {
+	t.Parallel()
+
+	log := slog.Default()
+
+	type state struct {
+		existingTokens []*apiv2.Token
+		adminSubjects  []string
+		projectRoles   map[string]apiv2.ProjectRole
+		tenantRoles    map[string]apiv2.TenantRole
+	}
+	tests := []struct {
+		name           string
+		sessionToken   *apiv2.Token
+		req            *adminv2.TokenServiceListRequest
+		state          state
+		wantErr        bool
+		wantErrMessage string
+		want           []*apiv2.Token
+	}{
+		{
+			name: "no tokens",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req:   &adminv2.TokenServiceListRequest{},
+			state: state{},
+			want:  nil,
+		},
+		{
+			name: "list tokens",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &adminv2.TokenServiceListRequest{},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+					User: "not phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+		},
+		{
+			name: "query uuid",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &adminv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					Uuid: new("c223af4d-b3f5-4df6-8815-52b80323930d"),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+		},
+		{
+			name: "query description and labels",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &adminv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					Labels: &apiv2.Labels{
+						Labels: map[string]string{
+							"a": "b",
+							"c": "d",
+						},
+					},
+					Description: new("test"),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid:        "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User:        "phippy",
+						Description: "test",
+						Meta: &apiv2.Meta{
+							Labels: &apiv2.Labels{
+								Labels: map[string]string{
+									"c": "d",
+									"a": "b",
+								},
+							},
+						},
+					},
+					{
+						Uuid:        "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User:        "phippy",
+						Description: "nope",
+						Meta: &apiv2.Meta{
+							Labels: &apiv2.Labels{
+								Labels: map[string]string{
+									"a": "b",
+									"c": "d",
+								},
+							},
+						},
+					},
+					{
+						Uuid:        "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User:        "phippy",
+						Description: "test",
+						Meta: &apiv2.Meta{
+							Labels: &apiv2.Labels{
+								Labels: map[string]string{
+									"a": "b",
+									"c": "nope",
+								},
+							},
+						},
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid:        "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User:        "phippy",
+					Description: "test",
+					Meta: &apiv2.Meta{
+						Labels: &apiv2.Labels{
+							Labels: map[string]string{
+								"a": "b",
+								"c": "d",
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "query user",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &adminv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					User: new("not phippy"),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+					User: "not phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+		},
+		{
+			name: "query token type",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &adminv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					TokenType: new(apiv2.TokenType_TOKEN_TYPE_API),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid:      "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User:      "phippy",
+						TokenType: apiv2.TokenType_TOKEN_TYPE_API,
+					},
+					{
+						Uuid:      "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User:      "phippy",
+						TokenType: apiv2.TokenType_TOKEN_TYPE_USER,
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid:      "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User:      "phippy",
+					TokenType: apiv2.TokenType_TOKEN_TYPE_API,
+					Meta:      &apiv2.Meta{},
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			ctx, cancel := context.WithCancel(token.ContextWithToken(t.Context(), tt.sessionToken))
+			defer cancel()
+
+			s := miniredis.RunT(t)
+			c := redis.NewClient(&redis.Options{Addr: s.Addr()})
+
+			tokenStore := token.NewRedisStore(c)
+			certStore := certs.NewRedisStore(&certs.Config{
+				RedisClient: c,
+			})
+
+			projectsAndTenantsGetter := func(ctx context.Context, userId string) (*api.ProjectsAndTenants, error) {
+				return &api.ProjectsAndTenants{
+					ProjectRoles: tt.state.projectRoles,
+					TenantRoles:  tt.state.tenantRoles,
+				}, nil
+			}
+
+			for _, tok := range tt.state.existingTokens {
+				require.NoError(t, tokenStore.Set(t.Context(), tok))
+			}
+
+			service := tokenService{
+				tokenCreator: *token.NewWithPermissionCheck(&token.TokenWithPermissionCheckConfig{
+					TokenWithoutPermissionCheckConfig: token.TokenWithoutPermissionCheckConfig{
+						Certs:  certStore,
+						Tokens: tokenStore,
+						Issuer: "",
+					},
+					Log:                      log,
+					AdminSubjects:            []string{},
+					Authorizer:               nil,
+					ProjectsAndTenantsGetter: projectsAndTenantsGetter,
+				}),
+				tokenStore: tokenStore,
+			}
+
+			if tt.wantErr == false {
+				// Execute proto based validation
+				err := protovalidate.Validate(tt.req)
+				require.NoError(t, err)
+			}
+
+			response, err := service.List(ctx, tt.req)
+
+			switch {
+			case tt.wantErr && err != nil:
+				if diff := cmp.Diff(tt.wantErrMessage, err.Error()); diff != "" {
+					t.Errorf("diff = %s", diff)
+				}
+
+			case tt.wantErr && err == nil:
+				t.Fatalf("want error %q, got response %q", tt.wantErrMessage, response)
+			case err != nil:
+				t.Fatalf("want response, got error %q", err)
+
+			default:
+				if diff := cmp.Diff(
+					tt.want, response.Tokens,
+					protocmp.Transform(),
+					protocmp.IgnoreFields(
+						&apiv2.Token{}, "issued_at", "uuid", "expires",
+					),
+					protocmp.IgnoreFields(
+						&apiv2.Meta{}, "created_at", "updated_at",
+					),
+				); diff != "" {
+					t.Errorf("diff: %s", diff)
+				}
+			}
+		})
+	}
+}
