+++ pkg/service/api/token/token-service_test.go
@@ -2,9 +2,7 @@ package token
 
 import (
 	"context"
-	"errors"
 	"log/slog"
-	"strings"
 	"testing"
 	"time"
 
@@ -20,7 +18,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/testing/protocmp"
-	"google.golang.org/protobuf/types/known/durationpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
@@ -28,75 +25,6 @@ var (
 	kubies = "00000000-0000-0000-0000-000000000000"
 	token1 = "00000000-0000-0000-0000-000000000000"
 )
-
-func Test_tokenService_CreateConsoleTokenWithoutPermissionCheck(t *testing.T) {
-	t.Parallel()
-	ctx := t.Context()
-	s := miniredis.RunT(t)
-	c := redis.NewClient(&redis.Options{Addr: s.Addr()})
-
-	tokenStore := token.NewRedisStore(c)
-	certStore := certs.NewRedisStore(&certs.Config{
-		RedisClient: c,
-	})
-
-	service := New(Config{
-		Log:        slog.Default(),
-		TokenStore: tokenStore,
-		CertStore:  certStore,
-		Issuer:     "http://test",
-	})
-
-	got, err := service.CreateUserTokenWithoutPermissionCheck(ctx, "test", new(1*time.Minute))
-	require.NoError(t, err)
-	// verifying response
-
-	require.NotNil(t, got)
-	require.NotNil(t, got)
-	require.NotNil(t, got.GetToken())
-
-	assert.NotEmpty(t, got.GetSecret())
-	assert.True(t, strings.HasPrefix(got.GetSecret(), "ey"), "not a valid jwt token") // jwt always starts with "ey" because it's b64 encoded JSON
-	claims, err := token.ParseJWTToken(got.GetSecret())
-	require.NoError(t, err, "token claims not parsable")
-	require.NotNil(t, claims)
-
-	assert.NotEmpty(t, got.GetToken().GetUuid())
-	assert.Equal(t, "test", got.GetToken().GetUser())
-
-	// verifying keydb entry
-	err = tokenStore.Set(ctx, got.GetToken())
-	require.NoError(t, err)
-
-	// listing tokens
-
-	tokenList, err := service.List(token.ContextWithToken(ctx, got.Token), &apiv2.TokenServiceListRequest{})
-	require.NoError(t, err)
-
-	require.NotNil(t, tokenList)
-	require.NotNil(t, tokenList)
-
-	require.Len(t, tokenList.Tokens, 1)
-
-	// Check still present
-	_, err = tokenStore.Get(ctx, got.GetToken().GetUser(), got.GetToken().GetUuid())
-	require.NoError(t, err)
-
-	// Check unpresent after revocation
-	err = tokenStore.Revoke(ctx, got.GetToken().GetUser(), got.GetToken().GetUuid())
-	require.NoError(t, err)
-
-	_, err = tokenStore.Get(ctx, got.GetToken().GetUser(), got.GetToken().GetUuid())
-	require.Error(t, err)
-
-	// List must now be empty
-	tokenList, err = service.List(token.ContextWithToken(ctx, got.Token), &apiv2.TokenServiceListRequest{})
-	require.NoError(t, err)
-
-	require.NotNil(t, tokenList)
-	require.NotNil(t, tokenList)
-	require.Empty(t, tokenList.Tokens)
-}
 
 func Test_Create(t *testing.T) {
 	t.Parallel()
@@ -507,7 +435,17 @@ func Test_Create(t *testing.T) {
 				issuer:                   "http://test",
 				adminSubjects:            tt.state.adminSubjects,
 				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+				tokenCreator: token.NewWithPermissionCheck(&token.TokenWithPermissionCheckConfig{
+					TokenWithoutPermissionCheckConfig: token.TokenWithoutPermissionCheckConfig{
+						Certs:  certStore,
+						Tokens: tokenStore,
+						Issuer: "http://test",
+					},
+					Log:                      log,
+					AdminSubjects:            tt.state.adminSubjects,
+					Authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+					ProjectsAndTenantsGetter: projectsAndTenantsGetter,
+				}),
 			}
 
 			if tt.wantErr == false {
@@ -542,676 +480,11 @@ func Test_Create(t *testing.T) {
 						&apiv2.Token{}, "issued_at", "uuid", "expires",
 					),
 					protocmp.IgnoreFields(
-						&apiv2.Meta{}, "created_at",
+						&apiv2.Meta{}, "created_at", "updated_at",
 					),
 				); diff != "" {
 					t.Errorf("diff: %s", diff)
 				}
-			}
-		})
-	}
-}
-
-func Test_CreateForUser(t *testing.T) {
-	t.Parallel()
-	type state struct {
-		adminSubjects []string
-		projectRoles  map[string]apiv2.ProjectRole
-		tenantRoles   map[string]apiv2.TenantRole
-	}
-	tests := []struct {
-		name           string
-		sessionToken   *apiv2.Token
-		req            *apiv2.TokenServiceCreateRequest
-		user           *string
-		state          state
-		wantErr        bool
-		wantErrMessage string
-		wantToken      *apiv2.Token
-	}{
-		{
-			name: "phippy can create token for user foo",
-			sessionToken: &apiv2.Token{
-				User:         "phippy",
-				Permissions:  []*apiv2.MethodPermission{},
-				ProjectRoles: map[string]apiv2.ProjectRole{},
-				TenantRoles:  map[string]apiv2.TenantRole{},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "empty token",
-			},
-			user: new("foo"),
-			state: state{
-				adminSubjects: []string{"phippy"},
-			},
-			wantToken: &apiv2.Token{
-				User:        "foo",
-				Description: "empty token",
-				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
-				Meta:        &apiv2.Meta{},
-			},
-		},
-		{
-			name: "bar can not create token for user foo",
-			sessionToken: &apiv2.Token{
-				User:         "bar",
-				Permissions:  []*apiv2.MethodPermission{},
-				ProjectRoles: map[string]apiv2.ProjectRole{},
-				TenantRoles:  map[string]apiv2.TenantRole{},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "empty token",
-			},
-			user: new("foo"),
-			state: state{
-				adminSubjects: []string{"phippy"},
-			},
-			wantToken:      nil,
-			wantErr:        true,
-			wantErrMessage: "permission_denied: only admins can specify token user",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx, cancel := context.WithCancel(token.ContextWithToken(t.Context(), tt.sessionToken))
-			defer cancel()
-
-			s := miniredis.RunT(t)
-			c := redis.NewClient(&redis.Options{Addr: s.Addr()})
-
-			tokenStore := token.NewRedisStore(c)
-			certStore := certs.NewRedisStore(&certs.Config{
-				RedisClient: c,
-			})
-
-			projectsAndTenantsGetter := func(ctx context.Context, userId string) (*api.ProjectsAndTenants, error) {
-				return &api.ProjectsAndTenants{
-					ProjectRoles: tt.state.projectRoles,
-					TenantRoles:  tt.state.tenantRoles,
-				}, nil
-			}
-			log := slog.Default()
-			service := tokenService{
-				log:                      log,
-				tokens:                   tokenStore,
-				certs:                    certStore,
-				issuer:                   "http://test",
-				adminSubjects:            tt.state.adminSubjects,
-				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
-			}
-
-			if tt.wantErr == false {
-				// Execute proto based validation
-				err := protovalidate.Validate(tt.req)
-				require.NoError(t, err)
-			}
-
-			response, err := service.CreateTokenForUser(ctx, tt.user, tt.req)
-			switch {
-			case tt.wantErr && err != nil:
-				if dff := cmp.Diff(tt.wantErrMessage, err.Error()); dff != "" {
-					t.Fatal(dff)
-				}
-			case tt.wantErr && err == nil:
-				t.Fatalf("want error %q, got response %q", tt.wantErrMessage, response)
-			case err != nil:
-				t.Fatalf("want response, got error %q", err)
-
-			default:
-				if response.Secret == "" {
-					t.Error("response secret for token may not be empty")
-				}
-				require.NotNil(t, tt.wantToken, "token returned, nil expected")
-
-				got := response.Token
-
-				if diff := cmp.Diff(
-					tt.wantToken, got,
-					protocmp.Transform(),
-					protocmp.IgnoreFields(
-						&apiv2.Token{}, "issued_at", "uuid", "expires",
-					),
-					protocmp.IgnoreFields(
-						&apiv2.Meta{}, "created_at",
-					),
-				); diff != "" {
-					t.Errorf("diff: %s", diff)
-				}
-			}
-		})
-	}
-}
-
-func Test_validateTokenRequest(t *testing.T) {
-	t.Parallel()
-	inOneHour := durationpb.New(time.Hour)
-	tests := []struct {
-		name          string
-		pat           *api.ProjectsAndTenants
-		token         *apiv2.Token
-		req           *apiv2.TokenServiceCreateRequest
-		adminSubjects []string
-		wantErr       error
-	}{
-		{
-			name: "simple token with empty permissions and roles",
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "",
-						Methods: []string{""},
-					},
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i don't need any permissions",
-				Expires:     inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       nil,
-		},
-		// Inherited Permissions
-		{
-			name: "simple token with no permissions but project role",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"ae8d2493-41ec-4efd-bbb4-81085b20b6fe": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"ae8d2493-41ec-4efd-bbb4-81085b20b6fe": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster for this project",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "ae8d2493-41ec-4efd-bbb4-81085b20b6fe",
-						Methods: []string{
-							"/metalstack.api.v2.IPService/Get",
-						},
-					},
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       nil,
-		},
-		// Permissions from Token
-		{
-			name: "simple token with one project and permission",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       nil,
-		},
-		{
-			name: "simple token with unknown method",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.UnknownService/Get"},
-					},
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("unknown method \"/metalstack.api.v2.UnknownService/Get\""),
-		},
-		{
-			name: "simple token with one project and permission, wrong project given",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
-					"cde": apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "sfs",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "cde",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("method \"/metalstack.api.v2.IPService/Get\" is not allowed on subject \"cde\" with your current user permissions"),
-		},
-		{
-			name: "simple token with one project and permission, wrong message given",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "sfs",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to list clusters",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/List"},
-					},
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("the following method \"/metalstack.api.v2.IPService/List\" is not allowed on any of the requested subjects: [abc]"),
-		},
-		{
-			name: "simple token with one project and permission, wrong messages given",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "sfs",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{
-							"/metalstack.api.v2.IPService/Create",
-							"/metalstack.api.v2.IPService/Get",
-							"/metalstack.api.v2.IPService/Delete",
-						},
-					},
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get and list clusters",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{
-							"/metalstack.api.v2.IPService/Get",
-							"/metalstack.api.v2.IPService/List",
-						},
-					},
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("the following method \"/metalstack.api.v2.IPService/List\" is not allowed on any of the requested subjects: [abc]"),
-		},
-		// Roles from Token
-		{
-			name: "token has no role",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-				TenantRoles: map[string]apiv2.TenantRole{
-					"john@github": apiv2.TenantRole_TENANT_ROLE_OWNER,
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("the following method \"/metalstack.api.v2.AuditService/Get\" is not allowed"),
-		},
-		{
-			name: "token has to low role",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_VIEWER,
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("the following method \"/metalstack.api.v2.ProjectService/Create\" is not allowed"),
-		},
-		{
-			name: "token request has unspecified role",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"abc": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_VIEWER,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_VIEWER,
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "abc",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-				},
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_UNSPECIFIED,
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("requested tenant role: \"TENANT_ROLE_UNSPECIFIED\" is not allowed"),
-		},
-		// AdminSubjects
-		{
-			name:          "requested admin role but is not allowed",
-			adminSubjects: []string{},
-			pat: &api.ProjectsAndTenants{
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get admin access",
-				AdminRole:   new(apiv2.AdminRole_ADMIN_ROLE_VIEWER),
-				Expires:     inOneHour,
-			},
-			wantErr: errors.New("the following method \"/grpc.reflection.v1.ServerReflection/ServerReflectionInfo\" is not allowed on any of the requested subjects: [*]"),
-		},
-		{
-			name: "requested admin role but is only viewer of admin orga",
-			adminSubjects: []string{
-				"company-a@github",
-			},
-			pat: &api.ProjectsAndTenants{
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_VIEWER,
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get admin access",
-				AdminRole:   new(apiv2.AdminRole_ADMIN_ROLE_EDITOR),
-				Expires:     inOneHour,
-			},
-			wantErr: errors.New("the following method \"/grpc.reflection.v1.ServerReflection/ServerReflectionInfo\" is not allowed on any of the requested subjects: [*]"),
-		},
-		{
-			name: "token requested admin role but is editor in admin orga",
-			adminSubjects: []string{
-				"company-a@github",
-			},
-			pat: &api.ProjectsAndTenants{
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "company-a@github",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get admin access",
-				AdminRole:   new(apiv2.AdminRole_ADMIN_ROLE_EDITOR),
-				Expires:     inOneHour,
-			},
-			wantErr: errors.New("the following method \"/grpc.reflection.v1.ServerReflection/ServerReflectionInfo\" is not allowed on any of the requested subjects: [*]"),
-		},
-		{
-			name: "token requested admin role and has admin role editor",
-			adminSubjects: []string{
-				"company-a@github",
-			},
-			pat: &api.ProjectsAndTenants{
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "company-a@github",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				TenantRoles: map[string]apiv2.TenantRole{
-					"company-a@github": apiv2.TenantRole_TENANT_ROLE_EDITOR,
-				},
-				AdminRole: new(apiv2.AdminRole_ADMIN_ROLE_EDITOR),
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get admin access",
-				AdminRole:   new(apiv2.AdminRole_ADMIN_ROLE_EDITOR),
-				Expires:     inOneHour,
-			},
-			wantErr: nil,
-		},
-		// Infra Roles
-		{
-			name: "admin editor requested infra editor",
-			adminSubjects: []string{
-				"company-admin@github",
-			},
-			token: &apiv2.Token{
-				User:      "company-admin@github",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				AdminRole: apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "metal-bmc token",
-				InfraRole:   apiv2.InfraRole_INFRA_ROLE_EDITOR.Enum(),
-				Expires:     inOneHour,
-			},
-			wantErr: nil,
-		},
-		{
-			name: "admin viewer requested infra editor",
-			adminSubjects: []string{
-				"company-admin@github",
-			},
-			token: &apiv2.Token{
-				User:      "company-admin@github",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				AdminRole: apiv2.AdminRole_ADMIN_ROLE_VIEWER.Enum(),
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "metal-bmc token",
-				InfraRole:   apiv2.InfraRole_INFRA_ROLE_EDITOR.Enum(),
-				Expires:     inOneHour,
-			},
-			wantErr: errors.New("the following method \"/metalstack.infra.v2.BMCService/BMCCommandDone\" is not allowed on any of the requested subjects: [*]"),
-		},
-		// Mixed role and permissions
-		{
-			name: "token has no role",
-			pat: &api.ProjectsAndTenants{
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"ae8d2493-41ec-4efd-bbb4-81085b20b6fe": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-			},
-			token: &apiv2.Token{
-				User:      "test",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
-				ProjectRoles: map[string]apiv2.ProjectRole{
-					"ae8d2493-41ec-4efd-bbb4-81085b20b6fe": apiv2.ProjectRole_PROJECT_ROLE_OWNER,
-				},
-			},
-			req: &apiv2.TokenServiceCreateRequest{
-				Description: "i want to get a cluster",
-				Permissions: []*apiv2.MethodPermission{
-					{
-						Subject: "ae8d2493-41ec-4efd-bbb4-81085b20b6fe",
-						Methods: []string{"/metalstack.api.v2.IPService/Get"},
-					},
-					{
-						Subject: "internet",
-						Methods: []string{"/metalstack.admin.v2.NetworkService/Create"},
-					},
-				},
-				TenantRoles: map[string]apiv2.TenantRole{
-					"john@github": apiv2.TenantRole_TENANT_ROLE_OWNER,
-				},
-				Expires: inOneHour,
-			},
-			adminSubjects: []string{},
-			wantErr:       errors.New("the following method \"/metalstack.admin.v2.NetworkService/Create\" is not allowed on any of the requested subjects: [internet]"),
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.token.User == "" {
-				t.Errorf("no user in token specified")
-			}
-
-			if tt.wantErr == nil {
-				// Execute proto based validation
-				err := protovalidate.Validate(tt.req)
-				require.NoError(t, err)
-			}
-
-			projectsAndTenantsGetter := func(ctx context.Context, userId string) (*api.ProjectsAndTenants, error) {
-				return tt.pat, nil
-			}
-			log := slog.Default()
-			service := tokenService{
-				log:                      log,
-				tokens:                   nil,
-				certs:                    nil,
-				issuer:                   "http://test",
-				adminSubjects:            tt.adminSubjects,
-				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
-			}
-
-			gotErr := service.validateTokenRequest(t.Context(), tt.token, tt.req)
-
-			t.Log(gotErr)
-			if tt.wantErr != nil {
-				require.EqualError(t, gotErr, tt.wantErr.Error())
-			} else if gotErr != nil {
-				require.NoError(t, gotErr)
 			}
 		})
 	}
@@ -1646,7 +919,17 @@ func Test_Update(t *testing.T) {
 				issuer:                   "http://test",
 				adminSubjects:            tt.state.adminSubjects,
 				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+				tokenCreator: token.NewWithPermissionCheck(&token.TokenWithPermissionCheckConfig{
+					TokenWithoutPermissionCheckConfig: token.TokenWithoutPermissionCheckConfig{
+						Certs:  certStore,
+						Tokens: tokenStore,
+						Issuer: "http://test",
+					},
+					Log:                      log,
+					AdminSubjects:            tt.state.adminSubjects,
+					Authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+					ProjectsAndTenantsGetter: projectsAndTenantsGetter,
+				}),
 			}
 
 			if tt.wantErr == false {
@@ -1786,7 +1069,17 @@ func Test_Refresh(t *testing.T) {
 				issuer:                   "http://test",
 				adminSubjects:            tt.state.adminSubjects,
 				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+				tokenCreator: token.NewWithPermissionCheck(&token.TokenWithPermissionCheckConfig{
+					TokenWithoutPermissionCheckConfig: token.TokenWithoutPermissionCheckConfig{
+						Certs:  certStore,
+						Tokens: tokenStore,
+						Issuer: "http://test",
+					},
+					Log:                      log,
+					AdminSubjects:            tt.state.adminSubjects,
+					Authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+					ProjectsAndTenantsGetter: projectsAndTenantsGetter,
+				}),
 			}
 
 			response, err := service.Refresh(ctx, &apiv2.TokenServiceRefreshRequest{})
@@ -2099,7 +1392,17 @@ func Test_List(t *testing.T) {
 				issuer:                   "http://test",
 				adminSubjects:            tt.state.adminSubjects,
 				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+				tokenCreator: token.NewWithPermissionCheck(&token.TokenWithPermissionCheckConfig{
+					TokenWithoutPermissionCheckConfig: token.TokenWithoutPermissionCheckConfig{
+						Certs:  certStore,
+						Tokens: tokenStore,
+						Issuer: "http://test",
+					},
+					Log:                      log,
+					AdminSubjects:            tt.state.adminSubjects,
+					Authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+					ProjectsAndTenantsGetter: projectsAndTenantsGetter,
+				}),
 			}
 
 			if tt.wantErr == false {
@@ -2129,7 +1432,7 @@ func Test_List(t *testing.T) {
 						&apiv2.Token{}, "issued_at", "uuid", "expires",
 					),
 					protocmp.IgnoreFields(
-						&apiv2.Meta{}, "created_at",
+						&apiv2.Meta{}, "created_at", "updated_at",
 					),
 				); diff != "" {
 					t.Errorf("diff: %s", diff)
